Move AliasRegistrar and PoliciesRegistrar
* with other registrars
