+++ st2common/st2common/bootstrap/aliasesregistrar.py
