+++ st2common/st2common/bootstrap/policiesregistrar.py
